feat: Create getters for the document and the heading schemas.
The getDocSchema is a finction that will create the document based on a boolean
If true no tag will render if false the article will.
The get heading Schema will generate a schema based on a boolean
If true only four heading can be allowed if not then all headings are allowed.